Split functionality.go into get_{header,payload}.go (#730)
